Refactored DDChart/Moved DDChart components
