+++ src/data/components/chart/Chart.tsx
@@ -1,0 +1,111 @@
+import React, { useState } from 'react'
+import {
+  GridItemHeader,
+  GridItemContent,
+  GridItem,
+} from '../../../components/GridItem'
+import DropdownPicker from '../../../components/DropdownPicker'
+import Line from './charts/Line'
+import Bar from './charts/Bar'
+import Pie from './charts/Pie'
+import Sunburst from './charts/Sunburst'
+import Radar from './charts/Radar'
+import BigChart from '../../../components/BigChart'
+import { ErrorText } from '../../../components/ErrorText'
+import { ChartType, ChartVariant, DDChartProps } from '../../types'
+import {
+  ChartDisplayOptions,
+  ChartVariantToggle,
+} from '../../../components/ChartDisplayOptions'
+import { ToggleBigChartButton } from '../../../components/ToggleBigChartButton'
+
+interface ChartGridItemProps {
+  isBig?: boolean
+  chartVariants: ChartVariant[]
+}
+
+const chartComponents: {
+  [key in ChartType]: (props: any) => JSX.Element
+} = {
+  Line: Line,
+  Bar: Bar,
+  Pie: Pie,
+  Radar: Radar,
+  Sunburst: Sunburst,
+}
+
+export default function Chart({
+  payload,
+  fullsize,
+  title,
+  description,
+  props,
+}: DDChartProps) {
+  const { setNames, sets } = payload as {
+    setNames: string[]
+    sets: { [key: string]: any }
+  }
+
+  const [set, setSet] = useState<string>(setNames?.[0] ?? '')
+  const [bigChartOpen, setBigChartOpen] = useState<boolean>(false)
+  const [chartVariantIdx, setChartVariantIdx] = useState<number>(0)
+
+  const { type: chartVariantToRender, props: chartProps } =
+    props.chartVariants[chartVariantIdx]
+  console.log(payload, chartVariantToRender, title)
+  const ChartComponent = chartComponents[chartVariantToRender]
+
+  const setNamesLength = payload.setNames ? payload.setNames.length : 0
+
+  const handleSetChange = (setName: string) => {
+    setSet(setName)
+  }
+
+  const ChartGridItem = ({
+    isBig = false,
+    chartVariants,
+  }: ChartGridItemProps) => (
+    <>
+      <GridItemHeader title={title} description={description} big={isBig}>
+        {setNamesLength > 1 ? (
+          <DropdownPicker
+            values={setNames}
+            onChange={handleSetChange}
+            selected={set}
+            big={isBig}
+          />
+        ) : null}
+      </GridItemHeader>
+      {sets ? (
+        <GridItemContent>
+          <ChartDisplayOptions>
+            {chartVariants.length > 1 ? (
+              <ChartVariantToggle
+                chartVariants={chartVariants}
+                selected={chartVariantIdx}
+                onChange={setChartVariantIdx}
+                big={isBig}
+              />
+            ) : null}
+            <ToggleBigChartButton
+              big={isBig}
+              onChange={() => setBigChartOpen(!bigChartOpen)}
+            />
+          </ChartDisplayOptions>
+          <ChartComponent big={isBig} data={sets[set]} {...chartProps} />
+        </GridItemContent>
+      ) : (
+        <ErrorText />
+      )}
+    </>
+  )
+
+  return (
+    <GridItem fullSize={fullsize}>
+      <ChartGridItem chartVariants={props.chartVariants} />
+      <BigChart open={bigChartOpen} onClose={() => setBigChartOpen(false)}>
+        <ChartGridItem isBig chartVariants={props.chartVariants} />
+      </BigChart>
+    </GridItem>
+  )
+}
