+++ src/data/components/chart/charts/Bar.tsx
