+++ src/data/components/chart/charts/common.tsx
